# docs/spec/dedi-global/openAPI.yaml
openapi: 3.0.3
info:
  title: DeDi API
  description: |
    DeDi (Decentralized Directory) API provides comprehensive functionality for managing 
    namespaces, registries, and records in a decentralized directory system.
    
    ## Authentication
    
    The API supports dual authentication methods:
    - **API Key Authentication**: Bearer token in Authorization header (recommended for production)
    - **Cookie Authentication**: Token cookie for browser integration
    
    All authentication endpoints now support password-based authentication with robust security requirements.
    
    ## Base Concepts
    
    - **Namespace**: Top-level organization unit for grouping related registries
    - **Registry**: Schema-defined container for records within a namespace  
    - **Record**: Individual data entries conforming to registry schema
    
    ## Key Features
    
    - **Password-based Authentication**: Secure login with enforced password complexity
    - **Domain Verification**: Verify ownership of domains for namespace binding
    - **Bulk Operations**: CSV import/export for large-scale data management
    - **Real-time Notifications**: Watch system for monitoring changes via webhooks
    - **Delegation Management**: Fine-grained access control and user permissions
    - **Version Control**: Complete history tracking for all resources
    - **Advanced Search**: Powerful search with nested field support
...
    
    ## Rate Limiting
    
...
    
    ## Production Readiness
...
servers:
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /dedi/lookup/{namespace}:
    get:
      tags:
        - Access
      security: []
      summary: Namespace lookup
      description: |
        Retrieves detailed information about a specific namespace including metadata and configuration.
      operationId: lookupNamespace
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
          description: Namespace ID or domain name
        - name: version_id
          in: query
          schema:
            type: string
          description: Specific version to retrieve
        - name: as_on
          in: query
          schema:
            type: string
            format: date
          description: Historical lookup date (YYYY-MM-DD)
      responses:
        '200':
          description: Namespace retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Namespace details retrieved successfully
                  data:
                    $ref: '#/components/schemas/Namespace'
        '404':
          description: Namespace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /dedi/lookup/{namespace}/{registry_name}:
    get:
      tags:
        - Access
      security: []
      summary: Registry lookup
      description: |
        Retrieves detailed information about a specific registry including schema and configuration.
      operationId: lookupRegistry
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
          description: Namespace ID or domain name
        - name: registry_name
          in: path
          required: true
          schema:
            type: string
          description: Registry name
        - name: version_id
          in: query
          schema:
            type: string
          description: Specific version to retrieve
        - name: as_on
          in: query
          schema:
            type: string
            format: date
          description: Historical lookup date (YYYY-MM-DD)
      responses:
        '200':
          description: Registry retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Resource retrieved successfully
                  data:
                    $ref: '#/components/schemas/Registry'
        '404':
          description: Registry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /dedi/lookup/{namespace}/{registry_name}/{record_name}:
    get:
      tags:
        - Access
      security: []
      summary: Record lookup
      description: |
        Retrieves detailed information about a specific record including data and metadata.
      operationId: lookupRecord
      parameters:
        - name: namespace
          in: path
          required: true
          schema:
            type: string
          description: Namespace ID or domain name
        - name: registry_name
          in: path
          required: true
          schema:
            type: string
          description: Registry name containing the record
        - name: record_name
          in: path
          required: true
          schema:
            type: string
          description: Record name to lookup
        - name: version_id
          in: query
          schema:
            type: string
          description: Specific version to retrieve
        - name: as_on
          in: query
          schema:
            type: string
            format: date
          description: Historical lookup date (YYYY-MM-DD)
      responses:
        '200':
          description: Record retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Resource retrieved successfully
                  data:
                    $ref: '#/components/schemas/Record'
        '404':
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Query endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: API key authentication using Bearer token
    CookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: Cookie-based authentication for browser integration
  schemas:
...
    Namespace:
      type: object
      properties:
        domain:
          type: string
          nullable: true
          description: Verified primary domain for the namespace
        namespace_id:
          type: string
          description: Unique namespace identifier
        name:
          type: string
          description: Namespace name
        description:
          type: string
          description: Namespace description
        digest:
          type: string
          description: Cryptographic digest of namespace data
        genesis:
          type: string
          format: date-time
          nullable: true
          description: Namespace genesis timestamp
        meta:
          type: object
          description: Additional metadata
        version:
          type: string
          description: Current version
        version_count:
          type: integer
          description: Total number of versions
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        created_by:
          type: string
          description: Creator user ID
        state:
          type: string
          enum: [live]
          description: Current namespace state
        is_verified:
          type: boolean
          description: Whether the namespace currently has an active verified domain
        registry_count:
          type: integer
          description: Number of live registries in the namespace
        ttl:
          type: integer
          description: Time to live in seconds
        proof:
          $ref: '#/components/schemas/NamespaceProof'
    Registry:
      type: object
      properties:
        namespace:
          type: string
          description: Namespace name or primary domain
        namespace_id:
          type: string
          description: Parent namespace ID
        registry_id:
          type: string
          description: Unique registry identifier
        registry_name:
          type: string
          description: Registry name
        description:
          type: string
          description: Registry description
        digest:
          type: string
          description: Cryptographic digest of registry data
        tag:
          type: string
          nullable: true
          description: Schema tag associated with the registry
        schema:
          type: object
          description: JSON Schema for record validation
        genesis:
          type: string
          format: date-time
          nullable: true
          description: Registry genesis timestamp
        meta:
          type: object
          description: Additional metadata
        version:
          type: string
          description: Current version
        version_count:
          type: integer
          description: Total number of versions
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        created_by:
          type: string
          description: Creator user ID
        state:
          type: string
          enum: [live, inactive]
          description: Current registry state
        record_count:
          type: integer
          description: Number of live records in the registry
        query_allowed:
          type: boolean
          description: Whether the registry is queryable
        ttl:
          type: integer
          description: Time to live in seconds
        proof:
          $ref: '#/components/schemas/RegistryProof'
    Record:
      type: object
      properties:
        namespace:
          type: string
          description: Parent namespace name or primary domain
        namespace_id:
          type: string
          description: Parent namespace ID
        registry_id:
          type: string
          description: Parent registry ID
        registry_name:
          type: string
          description: Parent registry name
        record_id:
          type: string
          description: Unique record identifier
        record_name:
          type: string
          description: Record name
        description:
          type: string
          description: Record description
        digest:
          type: string
          description: Cryptographic digest of record data
        schema:
          type: object
          description: Registry schema associated with the record
        details:
          type: object
          description: Record data conforming to registry schema
        meta:
          type: object
          description: Additional metadata
        version:
          type: string
          description: Current version
        version_count:
          type: integer
          description: Total number of versions
        genesis:
          type: string
          format: date-time
          nullable: true
          description: Record genesis timestamp
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        created_by:
          type: string
          description: Creator user ID
        state:
          type: string
          enum: [draft, live]
          description: Current record state
        valid_till:
          type: string
          format: date-time
          nullable: true
          description: Record expiration timestamp
        ttl:
          type: integer
          description: Time to live in seconds
        proof:
          $ref: '#/components/schemas/RecordProof'
...
    NamespaceProof:
      type: object
      properties:
        type:
          type: string
          example: DediNamespaceProof2026
        namespace_did:
          type: string
        creator_did:
          type: string
        digest:
          type: string
        network_genesis:
          type: string
    RegistryProof:
      type: object
      properties:
        type:
          type: string
          example: DediRegistryProof2026
        namespace_did:
          type: string
        registry_identifier:
          type: string
        creator_did:
          type: string
        digest:
          type: string
        network_genesis:
          type: string
    RecordProof:
      type: object
      properties:
        type:
          type: string
          example: DediRecordProof2026
        namespace_did:
          type: string
        registry_identifier:
          type: string
        record_identifier:
          type: string
        creator_did:
          type: string
        digest:
          type: string
        network_genesis:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: An error occurred
        error:
          type: string
          description: Detailed error information
          example: Invalid request parameters
        code:
          type: string
          description: Error code
          example: INVALID_REQUEST
...